feat - 게시판 작성자를 nickname으로 보이게끔 작성, 게시글 더미데이터 추가
